add version to modulepath, don't rm modulepath
This PR embeds the "version" of a module into the downloaded path,
version being either a tag or branch. So if your Terrafile has:

```
aws-cognito-pool:
  source: "https://github.com/lgallard/terraform-aws-cognito-user-pool.git"
  version: "0.9.4"
aws-cognito-pool:
  source: "https://github.com/lgallard/terraform-aws-cognito-user-pool.git"
  version: "master"
```

The repos will be cloned to `./vendor/modules/aws-cognito-pool/0.9.4/`
and `./vendor/modules/aws-cognito-pool/master`

The other change is that `./vendor/modules` is no longer removed on
every run. If a directory already exists, an ERROR log is printed, but
the clone process continues.  Removal of the directories is now the
users responsibility.